Refactored Octy Job module
- Dynamic octy job payloads
- Preventing Duplicate octy jobs based on meta data attributes.
- Synced octy job schedulers.
Using an ownership model to ensure singular instances handle jobs independently allowing multiple octy job services to be initialised. 
- Octy job failsafe. 
Jobs that have a status of 'processing' for more than 24 hours have are reset to be re-run. This handles cases where workers hang due to resource limits being exceeded and no response is return to octy job service, meaning the job would hang in a 'processing' state and never be re-run.